third_party/blink/web_tests/flag-specific/highdpi/fast/events/continuous-platform-wheelevent-in-scrolling-div-expected.txt

This is a testharness.js-based test.
FAIL This test checks the wheel delta value of wheel events, which should be the number of ticks multiplies the legacy mouse wheel tick multiplier. assert_equals: expected -226 but got -339
Harness: the test ran to completion.
